+++ src/lib/bars/series-bar-stacked.ts
@@ -1,10 +1,15 @@
 import { scaleBand, ScaleBand, ScaleContinuousNumeric, scaleLinear } from 'd3-scale';
 import { BaseType, Selection } from 'd3-selection';
-import { COLORS_CATEGORICAL, dataSeries, DataSeries } from '../core';
+import { COLORS_CATEGORICAL, DataSeriesGenerator } from '../core';
 import { Size } from '../core/utils';
 import { DataBar, JoinEvent, seriesBar } from './series-bar';
+import { DataSeriesBarGrouped } from './series-bar-grouped';
 
-export interface DataSeriesBarStackedCreation {
+export interface DataBarStacked extends DataBar {
+  stackIndex: number;
+}
+
+export interface DataSeriesBarStacked extends DataSeriesGenerator<DataBarStacked> {
   mainValues: any[];
   mainScale: ScaleBand<any>;
   innerPadding: number;
@@ -14,91 +19,76 @@ export interface DataSeriesBarStackedCreation {
   flipped: boolean;
 }
 
-export function dataSeriesBarStackedCreation(
-  data?: Partial<DataSeriesBarStackedCreation>
-): DataSeriesBarStackedCreation {
+export function dataSeriesBarStacked(data: Partial<DataSeriesBarStacked>): DataSeriesBarStacked {
   return {
-    mainValues: data?.mainValues || [],
+    mainValues: data.mainValues || [],
     mainScale:
-      data?.mainScale ||
+      data.mainScale ||
       scaleBand()
-        .domain(data?.mainValues || [])
+        .domain(data.mainValues || [])
         .padding(0.1),
-    crossValues: data?.crossValues || [],
+    crossValues: data.crossValues || [],
     crossScale:
-      data?.crossScale ||
+      data.crossScale ||
       scaleLinear()
         .domain([
           0,
-          Math.max(...(data?.crossValues?.map((values) => values.reduce((a, b) => a + b)) || [])),
+          Math.max(...(data.crossValues?.map((values) => values.reduce((a, b) => a + b)) || [])),
         ])
         .nice(),
-    flipped: data?.flipped || false,
-    innerPadding: data?.innerPadding || 0.1,
+    flipped: data.flipped || false,
+    innerPadding: data.innerPadding || 0.1,
+    dataGenerator: data.dataGenerator || dataBarStackedGenerator,
   };
 }
 
-export interface DataBarStacked extends DataBar {
-  stackIndex: number;
-}
-
-export function dataBarsStacked(
-  creationData: DataSeriesBarStackedCreation,
-  bounds: Size
+export function dataBarStackedGenerator(
+  selection: Selection<Element, DataSeriesBarStacked>
 ): DataBarStacked[] {
-  if (!creationData.flipped) {
-    creationData.mainScale.range([0, bounds.width]);
-    creationData.crossScale.range([bounds.height, 0]);
+  const seriesDatum = selection.datum(),
+    bounds = selection.bounds()!;
+  if (!seriesDatum.flipped) {
+    seriesDatum.mainScale.range([0, bounds.width]);
+    seriesDatum.crossScale.range([bounds.height, 0]);
   } else {
-    creationData.mainScale.range([0, bounds.height]);
-    creationData.crossScale.range([0, bounds.width]);
+    seriesDatum.mainScale.range([0, bounds.height]);
+    seriesDatum.crossScale.range([0, bounds.width]);
   }
 
   const data: DataBarStacked[] = [];
-  for (let i = 0; i < creationData.mainValues.length; ++i) {
-    const subcategoryValues = creationData.crossValues[i];
+  for (let i = 0; i < seriesDatum.mainValues.length; ++i) {
+    const subcategoryValues = seriesDatum.crossValues[i];
     let sum = 0;
     for (let j = 0; j < subcategoryValues.length; ++j) {
-      const c = creationData.mainValues[i];
+      const c = seriesDatum.mainValues[i];
       const v = subcategoryValues[j];
 
-      if (!creationData.flipped) {
+      if (!seriesDatum.flipped) {
         data.push({
           stackIndex: i,
           index: j,
-          key: creationData.keys?.[i][j] || `${i}/${j}`,
-          x: creationData.mainScale(c)!,
-          y: -sum + Math.min(creationData.crossScale(0)!, creationData.crossScale(v)!),
-          width: creationData.mainScale.bandwidth(),
-          height: Math.abs(creationData.crossScale(0)! - creationData.crossScale(v)!),
+          key: seriesDatum.keys?.[i][j] || `${i}/${j}`,
+          x: seriesDatum.mainScale(c)!,
+          y: -sum + Math.min(seriesDatum.crossScale(0)!, seriesDatum.crossScale(v)!),
+          width: seriesDatum.mainScale.bandwidth(),
+          height: Math.abs(seriesDatum.crossScale(0)! - seriesDatum.crossScale(v)!),
         });
         sum += data[data.length - 1].height;
       } else {
         data.push({
           stackIndex: i,
           index: j,
-          key: creationData.keys?.[i][j] || `${i}/${j}`,
-          x: sum + Math.min(creationData.crossScale(0)!, creationData.crossScale(v)!),
-          y: creationData.mainScale(c)!,
-          width: Math.abs(creationData.crossScale(0)! - creationData.crossScale(v)!),
-          height: creationData.mainScale.bandwidth(),
+          key: seriesDatum.keys?.[i][j] || `${i}/${j}`,
+          x: sum + Math.min(seriesDatum.crossScale(0)!, seriesDatum.crossScale(v)!),
+          y: seriesDatum.mainScale(c)!,
+          width: Math.abs(seriesDatum.crossScale(0)! - seriesDatum.crossScale(v)!),
+          height: seriesDatum.mainScale.bandwidth(),
         });
         sum += data[data.length - 1].width;
       }
     }
   }
   return data;
-}
-
-export interface DataSeriesBarStacked extends DataSeries<DataBarStacked> {}
-
-export function dataSeriesBarStacked(
-  creationData: DataSeriesBarStackedCreation
-): DataSeriesBarStacked {
-  return dataSeries({
-    data: (s) => dataBarsStacked(creationData, s.bounds()!),
-    key: (d) => d.key,
-  });
 }
 
 export function seriesBarStacked<