+++ src/lib/bars/series-label.ts
@@ -1,26 +1,22 @@
 import { easeCubicOut } from 'd3-ease';
 import { BaseType, select, Selection } from 'd3-selection';
-import { debug, nodeToString, Position, positionToTransformAttr } from '../core';
-import { dataSeries, DataSeries } from '../core/series';
+import {
+  DataSeriesGenerator,
+  debug,
+  nodeToString,
+  Position,
+  positionToTransformAttr,
+} from '../core';
 
 export interface DataLabel extends Position {
   text: string | number;
   // todo: add index property?
-  key?: string;
-}
-
-export interface DataSeriesLabel extends DataSeries<DataLabel> {}
-
-export function dataSeriesLabel(data?: Partial<DataSeriesLabel>): DataSeriesLabel {
-  return dataSeries({
-    data: data?.data,
-    key: data?.key || ((d, i) => d.key || i),
-  });
+  key: string;
 }
 
 export function seriesLabel<
   GElement extends Element,
-  Datum extends DataSeriesLabel,
+  Datum extends DataSeriesGenerator<DataLabel>,
   PElement extends BaseType,
   PDatum
 >(
@@ -49,7 +45,7 @@ export function seriesLabel<
 
 export function seriesLabelRender<
   GElement extends Element,
-  Datum extends DataSeriesLabel,
+  Datum extends DataSeriesGenerator<DataLabel>,
   PElement extends BaseType,
   PDatum
 >(
@@ -57,10 +53,10 @@ export function seriesLabelRender<
 ): Selection<GElement, Datum, PElement, PDatum> {
   return selection.each((d, i, g) => {
     debug(`render label series on ${nodeToString(g[i])}`);
-    const series = select(g[i]);
+    const series = select<GElement, Datum>(g[i]);
     series
       .selectAll<SVGTextElement, DataLabel>('text')
-      .data(d.data instanceof Function ? d.data(series) : d.data, d.key)
+      .data(d.dataGenerator(series), (d) => d.key)
       .join(
         (enter) =>
           enter