+++ src/lib/legend/legend-squares.ts
@@ -1,45 +1,51 @@
+import { Selection } from 'd3-selection';
 import { DataLegend, DataLegendItem } from './legend';
 
-export interface DataLegendSquares {
-  title: string;
+export interface DataLegendSquares extends DataLegend {
   labels: string[];
   colors: string | string[] | ((label: string) => string);
   sizes: string | string[] | ((label: string) => string);
 }
 
-export function dataLegendSquares(data: Partial<DataLegendSquares>): DataLegend {
+export function dataLegendSquares(data: Partial<DataLegendSquares>): DataLegendSquares {
   return {
+    labels: data.labels || [],
+    colors: data.colors || '#000000',
+    sizes: data.sizes || '0.7em',
     title: data.title || '',
-    data: (s) => {
-      const labels = data?.labels || [];
-
-      return labels.map((l, i) => {
-        const color =
-          typeof data.colors === 'string'
-            ? data.colors
-            : Array.isArray(data.colors)
-            ? data.colors[i]
-            : data.colors instanceof Function
-            ? data.colors(l)
-            : '#232323';
-        const size =
-          typeof data.sizes === 'string'
-            ? data.sizes
-            : Array.isArray(data.sizes)
-            ? data.sizes[i]
-            : data.sizes instanceof Function
-            ? data.sizes(l.toString())
-            : '1em';
-        return {
-          label: l,
-          symbolTag: 'rect',
-          symbolAttributes: [
-            ...['width', 'height'].map((name) => ({ name: name, value: size })),
-            { name: 'fill', value: color },
-          ],
-        };
-      });
-    },
-    key: (d) => d.label.toString(),
+    dataGenerator: data.dataGenerator || dataLegendItemSquareGenerator,
   };
+}
+
+export function dataLegendItemSquareGenerator(
+  selection: Selection<Element, DataLegendSquares>
+): DataLegendItem[] {
+  const seriesDatum = selection.datum();
+
+  return seriesDatum.labels.map((l, i) => {
+    const color =
+      typeof seriesDatum.colors === 'string'
+        ? seriesDatum.colors
+        : Array.isArray(seriesDatum.colors)
+        ? seriesDatum.colors[i]
+        : seriesDatum.colors instanceof Function
+        ? seriesDatum.colors(l)
+        : '#232323';
+    const size =
+      typeof seriesDatum.sizes === 'string'
+        ? seriesDatum.sizes
+        : Array.isArray(seriesDatum.sizes)
+        ? seriesDatum.sizes[i]
+        : seriesDatum.sizes instanceof Function
+        ? seriesDatum.sizes(l.toString())
+        : '1em';
+    return {
+      label: l,
+      symbolTag: 'rect',
+      symbolAttributes: [
+        ...['width', 'height'].map((name) => ({ name: name, value: size })),
+        { name: 'fill', value: color },
+      ],
+    };
+  });
 }